api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.26.1
dlib==19.24.2
pydantic==2.4.2
python-dotenv==1.0.0
pillow==10.1.0 